speedlimit_ahead based on gasbuttonstatus, better deceleration
based on constant deceleration the formula does not include time but just initial and final velocities and distance.

https://www.khanacademy.org/science/ap-physics-1/ap-one-dimensional-motion/motion-with-constant-acceleration/a/1d-kinematics-review-article

on eco the distance to start braking is 300m normal is 200m and sport at 100m with a maximum brake of -3m/s/s which is required to brake from 100km/hr to 50 km/hr in 100m. 

Also moved speed limit deceleration after deceleration for turn so make it the priority.

Corrected deceleration for turn deceleration which was too high because of missing average of self.v_cruise and v_curvature, causing a higher deceleration to be created than needed.